mod: rename tms_pixellock to tms_datashader
mod: rename tms_pixellock to tms_datashader in github workflow
