+++ .github/workflows/unittests.yml
@@ -1,7 +1,7 @@
 # This workflow will install Python dependencies, run tests and lint with a variety of Python versions
 # For more information see: https://help.github.com/actions/language-and-framework-guides/using-python-with-github-actions
 
-name: Test tms_pixellock
+name: Test tms_datashader
 
 on:
   push:
