+++ tests/test_pandas_util.py
@@ -1,7 +1,7 @@
 #!/usr/bin/env pytest
 import pandas as pd
 import pytest
-from tms_pixellock_api.helpers import pandas_util
+from tms_datashader_api.helpers import pandas_util
 
 
 def test_simplify_categories_not_categorical():
