+++ tms_datashader_api/helpers/drawing.py
